fixed SystemState python issues
Not quite sure why it works, but...
Adding SystemState* to the class_ declaration in robocup-py wrapper for SystemState
fixes the issues with passing it into python

A few other things:
PyQt uses different wrapper technologies than boost python, so passing a Qt
object from python to C++ through a wrapper function doesnt work like youd think
I changed the wrapper for SystemState.draw_line to take in a 3-tuple, then we
convert it to a QColor on this side.  Similarly with QString.

I also removed the debug statements that were in place for trying to get
SystemState stuff to work.

I commented out the call to draw_circle in mark.py because I was lazy
and didnt write the wrapper for that one yet - well need to write the wrapper
then uncomment the call


Former-commit-id: 86014da7e49a8e7898d006d8a2a014e003180faf