per-robot-model tuning params.  closes #96
Each robot has a RobotConfig object that contains parameters such as:
* pid values
* angular and translational velocity multipliers
* max chip and kick values

MotionControl reads from the relevant params

There are separate config values for Rev2008 and Rev2011.  Each robot
gets its RobotConfig object from the Processor which reads the model
number from the radio packet to determine which configs to associate it
with.


Former-commit-id: b070419fa711a1411a6b4244b0a8a80271a4681c